Reorganized folder, wrote raindrops, updated main loop
