+++ niceTimer.h
